quic: various improvements
- set sockopt to receive ECN
- set socktop to dontfrag
- vquic: provide ECN on receive to callbacks
- ngtcp2: manage sendbuf more carefully, do not add
  more than QUIC can keep in flight

+++ lib/vquic/cf-quiche.c
@@ -652,7 +652,7 @@ struct recv_ctx {
 static CURLcode cf_quiche_recv_pkts(const unsigned char *buf, size_t buflen,
                                     size_t gso_size,
                                     struct sockaddr_storage *remote_addr,
-                                    socklen_t remote_addrlen, int ecn,
+                                    socklen_t remote_addrlen, uint8_t ecn,
                                     void *userp)
 {
   struct recv_ctx *r = userp;
